ToDo/requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-filter
